+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 - New management command `eveuniverse_load_types` making it easier for apps to preload the eve objects they need
+- Type hints for all methods
+- Spinx docs
 
 ### Changed